# docs/api-specification.yaml
openapi: 3.0.1
info:
  title: PayFlow Payment Orchestration API
  description: "Multi-gateway payment orchestration (Razorpay, Stripe, PayU, UPI)\
    \ with intelligent routing, 2-second failover, idempotency, webhook reconciliation\
    \ and a full audit trail. Amounts are integer paise."
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- apiKey: []
...
paths:
...
  /api/v1/payments:
    get:
      tags:
      - Payments
      summary: Retrieve payments by merchant order id (#3)
      operationId: byOrder
      parameters:
      - name: merchant_order_id
        in: query
        required: true
        schema:
          maxLength: 128
          minLength: 0
          type: string
      - name: X-Merchant-Id
        in: header
        required: false
        schema:
          type: string
          default: default
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentView"
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
    post:
      tags:
      - Payments
      summary: Initiate a payment (#1)
      description: "Routes to the best gateway, fails over within 2 s, and auto-captures\
        \ unless capture_mode=MANUAL. Returns 201 when authorised/captured, 202 when\
        \ queued (UPI collect pending or every gateway throttled), 402 on decline.\
        \ A duplicate Idempotency-Key gets 409 while the first request is in flight\
        \ and the cached response afterwards."
      operationId: create
      parameters:
      - name: Idempotency-Key
        in: header
        description: "Client-generated key (UUID recommended), scoped to the merchant"
        required: true
        schema:
          maxLength: 255
          minLength: 0
          type: string
      - name: X-Merchant-Id
        in: header
        required: false
        schema:
          pattern: "^[A-Za-z0-9_-]{1,64}$"
          type: string
          default: default
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
  /api/v1/payments/{id}/void:
    post:
      tags:
      - Payments
      summary: "Void an authorisation, or release the remaining hold of a partial\
        \ capture (#5)"
      operationId: voidPayment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          maxLength: 255
          minLength: 0
          type: string
      - name: X-Merchant-Id
        in: header
        required: false
        schema:
          type: string
          default: default
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
  /api/v1/payments/{id}/refund:
    post:
      tags:
      - Payments
      summary: "Refund a captured or settled payment, fully or partially (#6)"
      operationId: refund
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: amount_paise
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          maxLength: 255
          minLength: 0
          type: string
      - name: X-Merchant-Id
        in: header
        required: false
        schema:
          type: string
          default: default
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
  /api/v1/payments/{id}/capture:
    post:
      tags:
      - Payments
      summary: "Capture an authorised payment, fully or partially (#4)"
      description: Omit amount_paise to capture the whole remaining hold. A partial
        capture leaves the payment PARTIALLY_CAPTURED with remaining_hold_paise available
        for another capture or a void.
      operationId: capture
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: amount_paise
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          maxLength: 255
          minLength: 0
          type: string
      - name: X-Merchant-Id
        in: header
        required: false
        schema:
          type: string
          default: default
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
...
components:
  schemas:
...
    PaymentView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchantId:
          type: string
        merchantOrderId:
          type: string
        amountPaise:
          type: integer
          format: int64
        amount:
          type: string
        currency:
          type: string
        paymentMethod:
          type: string
        captureMode:
          type: string
        upiFlow:
          type: string
        state:
          type: string
        gateway:
          type: string
        gatewayReference:
          type: string
        capturedPaise:
          type: integer
          format: int64
        remainingHoldPaise:
          type: integer
          format: int64
        releasedPaise:
          type: integer
          format: int64
        refundedPaise:
          type: integer
          format: int64
        attemptsMade:
          type: integer
          format: int32
        retryCount:
          type: integer
          format: int32
        nextRetryAt:
          type: string
          format: date-time
        failureCode:
          type: string
        failureReason:
          type: string
        traceId:
          type: string
          format: uuid
        authorisedAt:
          type: string
          format: date-time
        authExpiresAt:
          type: string
          format: date-time
        settledAt:
          type: string
          format: date-time
        settlementBatchId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: PAYMENT_AUTH_FAILED
            message:
              type: string
              example: Payment authorisation was declined by the issuing bank.
            details:
              type: object
              example: "{gateway_error_code=BAD_REQUEST_ERROR, suggestion=Please try\
                \ a different payment method or contact your bank., gateway=razorpay,\
                \ gateway_error_description=The card issuer has declined this transaction.}"
            request_id:
              type: string
              example: req_a1b2c3d4e5f6
            trace_id:
              type: string
              example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
            timestamp:
              type: string
              example: 2025-03-19T10:30:00.000Z
      description: A7.2 standard error format
  securitySchemes:
    apiKey:
      type: apiKey
      name: X-API-Key
      in: header